Refactor out image layout logic into a trait for Video and Image asset types
