+++ app/src/main/scala/com/waz/zclient/messages/parts/AssetController.scala
@@ -18,6 +18,7 @@
 package com.waz.zclient.messages.parts
 
 import android.view.{View, ViewGroup}
+import android.widget.{LinearLayout, TextView}
 import com.waz.ZLog.ImplicitTag._
 import com.waz.ZLog._
 import com.waz.api
@@ -33,8 +34,10 @@ import com.waz.utils.events.Signal
 import com.waz.zclient._
 import com.waz.zclient.messages.MessageViewPart
 import com.waz.zclient.messages.parts.DeliveryState.{Complete, OtherUploading}
+import com.waz.zclient.utils.ContextUtils._
 import com.waz.zclient.utils.{RichView, StringUtils}
-import com.waz.zclient.views.{AssetActionButtonNew, AssetBackground}
+import com.waz.zclient.views.ImageController.WireImage
+import com.waz.zclient.views.{ImageAssetDrawable, AssetActionButtonNew, AssetBackground}
 import org.threeten.bp.Duration
 
 import scala.PartialFunction._
@@ -89,7 +92,7 @@ class AssetController(implicit inj: Injector) extends Injectable {
 
 }
 
-trait AssetPart extends ViewGroup with MessageViewPart with ViewHelper {
+trait AssetPart extends View with MessageViewPart with ViewHelper {
   val zms = inject[Signal[ZMessaging]]
   val assets = inject[AssetController]
   val message = Signal[MessageData]()
@@ -101,6 +104,13 @@ trait AssetPart extends ViewGroup with MessageViewPart with ViewHelper {
   val progressDots = new AssetBackground(deliveryState.map { case OtherUploading => true; case _ => false })
   setBackground(progressDots)
 
+  override def set(pos: Int, msg: MessageData, part: Option[MessageContent], widthHint: Int): Unit = {
+    message ! msg
+  }
+
+}
+
+trait ContentAssetPart extends AssetPart {
   def inflate(): Unit
 
   inflate()
@@ -111,13 +121,42 @@ trait AssetPart extends ViewGroup with MessageViewPart with ViewHelper {
     case OtherUploading => false
     case _ => true
   }.on(Threading.Ui)(content.setVisible)
+}
+
+trait ImageLayoutAssetPart extends AssetPart {
+  protected val imageDim = message map { _.imageDimensions.getOrElse(Dim2(1, 1)) }
+  protected val width = Signal[Int]()
+
+  val imageDrawable = new ImageAssetDrawable(message map { m => WireImage(m.assetId) })
+
+  val height = for {
+    w <- width
+    Dim2(imW, imH) <- imageDim
+  } yield imH * w / imW  // TODO: improve view size computation
+
+  height { h =>
+    val w = getLayoutParams.width
+    val margin = if (h > w) getDimenPx(R.dimen.content__padding_left) else 0
+    val displayWidth = w - 2 * margin
+    val height = (h * (displayWidth.toDouble / w)).toInt
+    //TODO, set the imageDrawable to draw only within the bounds of `displayWidth`
+    val pms = new LinearLayout.LayoutParams(w, height)
+    pms.setMargins(margin, 0, margin, 0)
+    setLayoutParams(pms)
+  }
 
   override def set(pos: Int, msg: MessageData, part: Option[MessageContent], widthHint: Int): Unit = {
-    message ! msg
+    super.set(pos, msg, part, widthHint)
+    width.mutateOrDefault(identity, widthHint)
+  }
+
+  override def onLayout(changed: Boolean, left: Int, top: Int, right: Int, bottom: Int): Unit = {
+    super.onLayout(changed, left, top, right, bottom)
+    width ! (right - left)
   }
 }
 
-trait ActionableAssetPart extends AssetPart {
+trait ActionableAssetPart extends ContentAssetPart {
   protected val assetActionButton: AssetActionButtonNew = findById(R.id.action_button)
 
   override def set(pos: Int, msg: MessageData, part: Option[MessageContent], widthHint: Int): Unit = {
@@ -132,6 +171,12 @@ trait PlayableAsset extends ActionableAssetPart {
     case _ => Duration.ZERO
   }
   val formattedDuration = duration.map(d => StringUtils.formatTimeSeconds(d.getSeconds))
+
+  protected val durationView: TextView = findById(R.id.duration)
+
+  //TODO there is more logic for what text to display in video views, but it doesn't seem to be used - confirm
+  formattedDuration.on(Threading.Ui)(durationView.setText)
+
 }
 
 sealed trait DeliveryState
