+++ app/src/main/scala/com/waz/zclient/messages/parts/VideoAssetPartView.scala
@@ -19,8 +19,7 @@ package com.waz.zclient.messages.parts
 
 import android.content.Context
 import android.util.AttributeSet
-import android.widget.{FrameLayout, LinearLayout, TextView}
-import com.waz.model.{Dim2, MessageContent, MessageData}
+import android.widget.FrameLayout
 import com.waz.threading.Threading
 import com.waz.zclient.R
 import com.waz.zclient.messages.MsgPart
@@ -29,21 +28,14 @@ import com.waz.zclient.utils.ContextUtils._
 import com.waz.zclient.utils.{AssetUtils, RichView}
 import com.waz.zclient.views.ImageAssetDrawable
 import com.waz.zclient.views.ImageAssetDrawable.State.Loaded
-import com.waz.zclient.views.ImageController.WireImage
 
-class VideoAssetPartView(context: Context, attrs: AttributeSet, style: Int) extends FrameLayout(context, attrs, style) with PlayableAsset {
+class VideoAssetPartView(context: Context, attrs: AttributeSet, style: Int) extends FrameLayout(context, attrs, style) with PlayableAsset with ImageLayoutAssetPart {
   def this(context: Context, attrs: AttributeSet) = this(context, attrs, 0)
-
   def this(context: Context) = this(context, null, 0)
 
   override val tpe: MsgPart = MsgPart.VideoAsset
 
   override def inflate() = inflate(R.layout.message_video_asset_content)
-
-  private val durationView: TextView = findById(R.id.duration)
-  inflate(R.layout.message_video_asset_content)
-
-  private val imageDrawable = new ImageAssetDrawable(assetId.map(WireImage))
 
   imageDrawable.state.map {
     case Loaded(_, _) => getColor(R.color.white)
@@ -59,9 +51,6 @@ class VideoAssetPartView(context: Context, attrs: AttributeSet, style: Int) exte
         case _ =>
       }
   }
-
-  //TODO there is more logic for what text to display, but it doesn't seem to be used - confirm
-  formattedDuration.on(Threading.Ui)(durationView.setText)
 
   actionReady.on(Threading.Ui) {
     case true =>
@@ -80,18 +69,5 @@ class VideoAssetPartView(context: Context, attrs: AttributeSet, style: Int) exte
         }
       }
     case _ =>
-  }
-
-  override def set(pos: Int, msg: MessageData, part: Option[MessageContent], widthHint: Int): Unit = {
-    super.set(pos, msg, part, widthHint)
-
-    val Dim2(w, h) = msg.imageDimensions.getOrElse(Dim2(1, 1))
-    val margin = if (h > w) getDimenPx(R.dimen.content__padding_left) else 0
-    val displayWidth = widthHint - 2 * margin
-    val height = (h * (displayWidth.toDouble / w)).toInt
-
-    val pms = new LinearLayout.LayoutParams(displayWidth, height)
-    pms.setMargins(margin, 0, margin, 0)
-    setLayoutParams(pms)
   }
 }
